refactor: fixing bug and change the logic feature

+++ src/pages/dashboard/management/home-villa/add/add-home-villa.tsx
@@ -7,7 +7,7 @@ import { useCreateApi, usePersistentData } from "../../../../../hooks";
 import { Layout } from "../../../../../components/ui";
 import { Button } from "../../../../../components";
 
-import { FaDownload, FaFirstdraft } from "react-icons/fa";
+import { FaDownload } from "react-icons/fa";
 
 import { General } from "./general";
 import { Media } from "./media";
@@ -17,13 +17,14 @@ import { VillaPolicies } from "./villa-policies";
 import { KeyFeatures } from "./key-features";
 
 import { Villa } from "../../../../../types";
+import { deleteCurrencyCode } from "../../../../../utils";
 
 const tabs = ["General", "Media", "Location", "Key Features", "Service & Features", "Villa Policies"];
 
 export const AddHomeVillaPage = () => {
   const { pathname } = useLocation();
 
-  const { mutate: addVillas } = useCreateApi("villas", ["villas"]);
+  const { mutate: addVillas } = useCreateApi<Partial<Villa>>("villas", ["add-villa"]);
 
   const useStore = usePersistentData<Partial<Villa>>("add-villa");
 
@@ -40,31 +41,36 @@ export const AddHomeVillaPage = () => {
 
     return () => {
       if (window.location.pathname !== "/dashboard/management/home-villa/add") {
+        if (!window.confirm("Are you sure you want to move the page before publish your villas?")) return;
         sessionStorage.clear();
       }
     };
   }, [activeTab, pathname]);
 
-  const handleSaveDraft = (e: React.MouseEvent) => {
-    e.preventDefault();
-    localStorage.setItem("data", data as string);
-  };
+  // const handleSaveDraft = (e: React.MouseEvent) => {
+  //   e.preventDefault();
+  //   setData(data);
+  // };
 
   const handlePublish = (e: React.MouseEvent) => {
     e.preventDefault();
-    addVillas(data);
+    const processData = deleteCurrencyCode(data);
+    addVillas(processData);
+    setTimeout(() => {
+      window.location.href = "/dashboard/management/home-villa";
+    }, 2000);
   };
 
   return (
     <Layout>
       {/* Header */}
-      <header className="flex items-center justify-between pb-4 mb-6 border-b border-dark/20">
+      <header className="flex items-center justify-between pb-4 mb-6 border-b border-dark/30">
         <h1 className="text-2xl font-bold">Add New Home & Villa</h1>
 
         <div className="flex items-center gap-4">
-          <Button className="flex items-center gap-2 btn-outline" onClick={handleSaveDraft}>
+          {/* <Button className="flex items-center gap-2 btn-outline" onClick={handleSaveDraft}>
             <FaFirstdraft /> Save as draft
-          </Button>
+          </Button> */}
           <Button className="flex items-center gap-2 btn-primary" onClick={handlePublish}>
             <FaDownload /> Publish
           </Button>
@@ -73,7 +79,7 @@ export const AddHomeVillaPage = () => {
 
       <div className="flex">
         {tabs.map((tab) => (
-          <button key={tab} className={`px-4 py-1.5 border border-dark/20 rounded-t-md ${activeTab === tab ? "bg-primary text-light" : "bg-light text-primary"}`} onClick={() => setActiveTab(tab)}>
+          <button key={tab} className={`px-4 py-1.5 border border-dark/30 rounded-t-md ${activeTab === tab ? "bg-primary text-light" : "bg-light text-primary"}`} onClick={() => setActiveTab(tab)}>
             {tab}
           </button>
         ))}
